Temporary remove deprecated attribute annotations
Since the whole state_service module is annotated as deprecated alaready, the deprecation warnings that stem from the fields we marked as deprecated on this PR get confused with the rest. Also, there were allow(deprecated) exceptions that canceled everything out, so:

- commented out deprecated annotation on StateService, StateServiceSubscriber and StateServiceError
- commented out allow(deprecated) on implementors of the above mentioned

this way we now get the right warnings just for the fields we are deprecating on this PR

one apparent problem is that state.rs being a massive 2.5k single file module, we get all warnings for all method implementations at once. It would be WAY cleaner if we had a file per implemented method, that way it would be overly clear which methods have been properly ported already.